requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
scikit-learn==1.5.1
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
imbalanced-learn==0.12.2
pydantic==2.7.1
